Added default profile pic, fixed placement of username
Previously if a profile picture was missing completely (even the default one),
then the username placement went to the top of the box below the missing
image placeholder text instead of staying at the bottom of the image box
like it should. I fixed this by putting the img element inside a div with
a set width and height so that even if the image is missing, the username
will go below that div in the right spot.