Remove sign-in req for viewing directory in navbar, show admin users
Previous template logic in profile_page.html would say there are no profiles
to display if even one user is an admin, which is a problem because there
will always be an admin user and hence message would always display. There is
probably a way to fix that, but instead I just allowed admin user's profiles
to display anyway because I don't think it's a big deal if they're shown.