Profile page: added tests and modifications
- Only display age instead of birth date on profile page for better privacy
- Set age to null (None) in view if user has not provided a birth date, and
  check if age exists in template rather than if birth date exists since that is
now checked in the view
- In the positive test case, I did not check if the user's profile picture is
  displayed because I am not sure how to do that and I do not think it's
worth it to figure it out now. I may come back to it later.
- I did not write negative test cases for non-existent users (either no username
  provided or no user with the provided username) because Django already throws
errors from the reverse() function. In future I may want to look into handling
and displaying those errors to the user on the front end, but it's not that
important for now.