Updated to Bootstrap 4, made pagination and image gallery changes
- Updated Bootstrap to v.4.
- Moved JS scripts to bottom of the base html page.
- Applied Bootstrap styling to image thumbnail gallery and pagination on Images
page.
- Removed uploader name under image thumbnails to simplify design. Therefore,
I also deleted the base and main image gallery templates and replaced them with
 a single image gallery template instead because the base template was just to
allow the optional inclusion of the uploader name below the image which the main
image gallery then added.
- Added first, last, and page number links to pagination on Images page. Wrote
function for determining how many and which page numbers to display and wrote
tests for it.
- Added bubbles around tags so they are more easily distinguishable. Will do
  more work on tags later as the styling and layout still isn't quite right.
- Removed unused styles from base.css.
- Updated some existing Image tests to take into account the new styling on
Images and tags.