Fixed image upload test for success case
The problem was that the
test_upload_images_page_redirects_to_login_page_when_not_logged_in_POST test was
posting the same test image, which moved the file pointer to the end of the
file. Then when test_upload_images_valid_image tried to post the file, it was
considered empty since the file pointer was at the end. The solution was to
reset the file pointer back to the beginning in tearDown() so each test that
uses the test image will read the whole file again.

Also, previously the upload valid image test was trying to check the filename of
the uploaded image against the test image, but since the filename is randomly
generated when uploaded this was not straightforward to test. Thus, I just
checked the upload by making sure no images existed before the upload and then
checking that one image exists after the upload since no other images are
uploaded in this test.