Fixed change profile picture test for success case
Same problem as other image uploading test: test image was being read
in another test which set the file pointer to the end of the file.
Solution was to reset the file pointer to beginning of file in tearDown().
Updated asserts: check that user does not have profile picture before upload
and then has one after upload.